+++ packages/core/src/components/ui/Toggle/Toggle.tsx
@@ -25,6 +25,9 @@ export interface ToggleProps
    * Specifies that this input should be disabled.
    */
   disabled?: boolean
+  /**
+   * Controls the component's direction.
+   */
   variant?: 'horizontal' | 'vertical'
 }